fix(audit): align schema pattern detection
